requirements.txt
# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0

# AI Orchestration
langgraph>=0.2.60
langchain-core>=0.3.28
langchain-google-genai>=2.0.7

# Google Cloud
google-cloud-bigquery>=3.27.0

# Data & Typing
pydantic>=2.11.0
pydantic-settings>=2.9.0

# Environment
python-dotenv>=1.0.1
